+++ tooling/support/android/src/main/kotlin/com/squareup/tooling/support/android/VariantExtractor.kt
@@ -53,7 +53,10 @@ internal fun BaseVariant.extractSquareVariantConfigurationParams(
   )
 
   // SqlDelight doesn't add src/*/sqldelight to the source sets, but changes to those files can affect other projects
-  if (project.plugins.hasPlugin("com.squareup.sqldelight")) {
+  if (
+    project.plugins.hasPlugin("com.squareup.sqldelight")
+      || project.plugins.hasPlugin("app.cash.sqldelight")
+  ) {
     val aliasFile = project.projectDir
       .resolve("src")
       .resolve(name)